requirements.txt
fastapi==0.125.0
langchain_chroma==1.1.0
langchain_classic==1.0.0
langchain_community==0.4.1
langchain_core==1.2.3
langchain_huggingface==1.2.0
langchain_ollama==1.0.1
langchain_text_splitters==1.1.0
pydantic==1.10.17
python-dotenv==1.2.1
